Split into oci_archive.rs and oci_dir.rs
